REBASED: Original repo 'npm-backup'. Improved build-chain and other updates.
In /src/:
	Formatted, linted, and spell-checked. Needs to be updated and tested
with newer versions of dependencies. The `OpsPipelineQueue` code should
be moved to a separate project, and either made a dependency or a git
sub-module if possible.

In /.husky/:
	Added git pre-commit hook to check linting and formatting, and run
test suite.

In /__tests__/:
	Formatted, linted, and spell-checked tests. Also updated some typings
for newer TS versions and added some better practices.

In TypeScript Config:
	Removed build-chain configuration.

In /:
	Cleaned up NPM config and updated packages. Updated formatting and
linting config. Spell-checked most files, and added cspell config.